zng_wgt_toggle

Function checked_popup

pub fn checked_popup(
    child: impl IntoUiNode,
    popup: impl IntoVar<WidgetFn<()>>,
) -> UiNode

P Popup open when the toggle button is checked.

This property can be used together with the ComboStyle! to implement a combo-box flyout widget.

The popup can be any widget, that will be open using POPUP, a Popup! or derived widget is recommended.

Note that if the checked property is not set the toggle will never be checked, to implement a drop-down menu set checked = var(false);.