Struct zng_wgt_toggle::Toggle

pub struct Toggle(/* private fields */);

W A toggle button that flips a bool or Option<bool> variable on click, or selects a value.

This widget has three primary properties, checked, checked_opt and value, setting one of the checked properties to a read-write variable enables the widget and it will set the variables on click, setting value turns the toggle in a selection item that is inserted/removed in a contextual selector.

Implementations

impl Toggle

pub fn widget_new() -> Self

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl Toggle

pub fn style_fn(&self, style_fn: impl IntoVar<StyleFn>)

P Extends or replaces the widget style.

Properties and when conditions in the style are applied to the widget. Style instances extend the contextual style by default, you can set replace on a style to true to fully replace.

impl Toggle

pub fn checked(&self, checked: impl IntoVar<bool>)

P Toggle cycles between true and false, updating the variable.

Note that you can read the checked state of the widget using is_checked.

impl Toggle

pub fn checked_opt(&self, checked: impl IntoVar<Option<bool>>)

P Toggle cycles between Some(true) and Some(false) and accepts None, if the widget is tristate also sets to None in the toggle cycle.

impl Toggle

pub fn tristate(&self, enabled: impl IntoVar<bool>)

P Enables None as an input value.

Note that None is always accepted in checked_opt, this property controls if None is one of the values in the toggle cycle. If the widget is bound to the checked property this config is ignored.

This is not enabled by default.

impl Toggle

pub fn is_checked(&self, state: impl IntoVar<bool>)

P If the toggle is checked from any of the three primary properties.

Note to read the tristate directly use IS_CHECKED_VAR directly.

impl Toggle

pub fn value<T: VarValue>(&self, value: impl IntoVar<T>)

P Values that is selected in the contextual selector.

The widget is_checked when the value is selected, on click and on value update, the selection is updated according to the behavior defined in the contextual selector. If no contextual selector is the widget is never checked.

Note that the value can be any type, but must be one of the types accepted by the contextual selector, type validation happens in run-time, an error is logged if the type is not compatible. Because any type can be used in this property type inference cannot resolve the type automatically and a type annotation is required: value<T> = t;.

This property interacts with the contextual selector, when the widget is clicked or the value variable changes the contextual Selector is used to implement the behavior.

impl Toggle

pub fn scroll_on_select(&self, enabled: impl IntoVar<bool>)

P If the scrolls into view when the value selected.

This is enabled by default.

impl Toggle

pub fn selector(&self, selector: impl IntoValue<Selector>)

P Sets the contextual selector that all inner widgets will target from the value property.

All value properties declared in widgets inside child will use the Selector to manipulate the selection.

Selection in a context can be blocked by setting the selector to Selector::nil(), this is also the default selector so the value property only works if a contextual selector is present.

This property sets the SELECTOR context and handles cmd::SelectOp requests. It also sets the widget access role to AccessRole::RadioGroup.

impl Toggle

pub fn select_on_init(&self, enabled: impl IntoVar<bool>)

P If value is selected when the widget that has the value is inited.

impl Toggle

pub fn deselect_on_deinit(&self, enabled: impl IntoVar<bool>)

P If value is deselected when the widget that has the value is deinited and the value was selected.

impl Toggle

pub fn select_on_new(&self, enabled: impl IntoVar<bool>)

P If value selects the new value when the variable changes and the previous value was selected.

impl Toggle

pub fn deselect_on_new(&self, enabled: impl IntoVar<bool>)

P If value deselects the previously selected value when the variable changes.

impl Toggle

pub fn checked_popup(&self, popup: impl IntoVar<WidgetFn<()>>)

P Popup open when the toggle button is checked.

This property can be used together with the ComboStyle! to implement a combo-box flyout widget.

The popup can be any widget, that will be open using POPUP, a Popup! or derived widget is recommended.

Note that if the checked property is not set the toggle will never be checked, to implement a drop-down menu set checked = var(false);.